Move price into component
